requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
djangorestframework==3.15.2
Faker==26.3.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2